EIP1-511 - Basic controller with 401/403 scenarios
