+++ src/test/kotlin/uk/gov/dluhc/printapi/rest/GetTemporaryCertificateSummariesByApplicationIdIntegrationTest.kt
@@ -1,0 +1,36 @@
+package uk.gov.dluhc.printapi.rest
+
+import org.junit.jupiter.api.Test
+import org.springframework.http.MediaType
+import uk.gov.dluhc.printapi.config.IntegrationTest
+import uk.gov.dluhc.printapi.testsupport.bearerToken
+import uk.gov.dluhc.printapi.testsupport.testdata.anotherValidEroId
+import uk.gov.dluhc.printapi.testsupport.testdata.getBearerToken
+
+internal class GetTemporaryCertificateSummariesByApplicationIdIntegrationTest : IntegrationTest() {
+
+    companion object {
+        private const val URI_TEMPLATE = "/eros/{ERO_ID}/temporary-certificates/applications/{APPLICATION_ID}"
+        private const val ERO_ID = "some-city-council"
+        private const val APPLICATION_ID = "7762ccac7c056046b75d4aa3"
+    }
+
+    @Test
+    fun `should return forbidden given user with valid bearer token belonging to a different ero`() {
+        wireMockService.stubCognitoJwtIssuerResponse()
+        val userGroupEroId = anotherValidEroId(ERO_ID)
+
+        webTestClient.get()
+            .uri(URI_TEMPLATE, ERO_ID, APPLICATION_ID)
+            .bearerToken(
+                getBearerToken(
+                    eroId = userGroupEroId,
+                    groups = listOf("ero-$userGroupEroId", "ero-vc-admin-$userGroupEroId")
+                )
+            )
+            .contentType(MediaType.APPLICATION_JSON)
+            .exchange()
+            .expectStatus()
+            .isForbidden
+    }
+}
